Delete dpkg call to avoid breakage on RPM-based systems
This commit gets rid of the unless condition in the call to install Ceph.
Running "ceph-deploy install" on a host where Ceph already is installed
doesn't cost a lot, but by having a hard-coded dpkg call in here, we're
in fact breaking compatibility with Red Hat and CentOS systems for no
real reason.