Fix handling of the file name of zapped status files
This commit changes the way that puppet-cephdeploy keeps track of which
disks have already been zapped. Previously, puppet-cephdeploy would
simply create a file for every device it has successfully zapped, and
that file would be inside of the "zap" directory of the ceph deploy
users home directory. Typically, device names are specified using a
"/dev/sda"-like syntax though, so infact, the modules would just try
to create $HOME/zap/dev/$NAME for every zapped disk. Because the name
may contain an arbitrary number of directories, this is a problem as
we can impossibly create matching subdirectories for every path that
may occur.

With this commit in place, the zap files will be created based on the
actual device ID, i.e. the last part of the device's specified path.