Configure the appropriate package repos on Debian/Ubuntu/RHEL/SLES
This commit adds functionality to puppet-cephdeploy to automatically
install the required repositories into the system configuration. The
inktank repositories are required to install ceph-deploy and the rest
of Ceph, so puppet-cephdeploy should be able to handle them. The
repos will automatically be deployed on hosts that have "cephdeploy"
assigned to them as a class.

The "ceph" and "ceph-extra" repositories will be added.

Currently, RHEL- and compatible systems as well as SLES 11 and the
Debian OS family (so Debian and Ubuntu) are supported.