Ensure that /lib/udev/rules.d/95-ceph-osd.rules works on RHEL
The udev on RHEL is too old to work with the rules version of this file
installed by upstream; upstream is thus providing a special version of
the file for RHEL-based distributions. The current emperor packages had
a bug in the package build spec file as documented on

https://github.com/ceph/ceph/pull/1149

Upstream has by now fixed the problem, but the fix didn't make it into
the emperor packages yet. We have to work around this problem in Puppet
by manually supplying a working version of the file. This commit adds
the according functionality.