Add a params class that automatically insert values from the environment
Currently, every class present in this needs to have parameters passed
individually, and there is no way to configure all parameters centrally.
That doesn't work well with the Puppet dashboard, as in the dashboard,
it is typical to create a group and classes and have the group and the
Dashboard classes parameters individually. With this patch, the classes
will try to fetch environment values for every individual parameter, so
also parameters defined via a Dashboard group will work.