Make setup_pools, has_compute and setup_virsh depend on "true"
Currently, the three if clauses for the mentioned values will trigger as
soon as the variables are set to any true value. Using the Dashboard,
any value set will count as true. Leaving the variables empty will lead
to them being set to "true" as well to imitate the previous default
behaviour. This change makes the parameters actually depend on "true" to
be triggered.