Run split on $disks if $disks is a string, not an array
The cephdeploy::osd function expects to receive an array containing
entries for the individual OSDs to be deployed. That doesn't work so
well with the Puppet dashboard because currently, the Puppet Dashboard
can only use strings as parameters. This commit adds a test for the
$disks variable; if it's a string, it will run split on the string and
create an array based on the split; if $disks is indeed an array, the
array will be passed on accordingly.