Ensure variable_prefix defaults to empty; add ceph_cluster_name
This commit ensures that the variable_prefix defaults to an empty value
or to the value of the configuration_variable_prefix parameter set by
the administrator in the site manifest or an ENC. It detaults to empty
to maintain compatibility with what the module previously did.

In addition, this commit adds a parameter for the ceph_cluster_name
parameter which was previously left out by accident.