Allow for separate Cinder and RBD CephX users and pools
Up to this point, the module has generally assumed that Glance and
Cinder would be using the client.admin ID to connect to Ceph. That
is possible a problem in terms of security, so the module should
allow the user to create a separate CephX user and a separate pool
for Cinder and RBD.

This commit adds the required functionality.