Ensure ceph-deploy gets installed on client nodes
This probably isn't the best solution, but as the yum.pp and apt.pp
classes have a "Before: ceph-deploy" clause in them, it'S the easiest
way to make the package installation mechanism work on clients like
it works on Ceph servers. The long-term solution for this would be to
actually change the yum and apt resources to depend on a resource that
does not enforce the installation of ceph-deploy.