Add an initial ceph.conf file to be used during the first mon creation
When creating the initial key for ceph.client.admin.keyring, ceph-deploy
will simply iterate through the list of servers. We don't have a way to
force it to use to a specific server right now.

When told to create the initial mon, ceph-deploy will report back that
is is done although ceph.client.admin.keyring may not be there yet. The
delay caused by this may cause the Cephdeploy puppet module to fail. At
the moment, we unfortunately don't have a way to know how long exactly
the delay might be.

We thus need to work around this and feed puppet-cephdeploy, during the
creating of the initial monitoring server, an initial ceph.conf which
only contains one monitoring server in the server list.